requirements.txt
Flask==2.3.3
supabase==2.3.0
python-dotenv==1.0.0